fix a few warnings in 32-bit builds (#66)

+++ samples/04_julia/main.cpp
@@ -164,7 +164,7 @@ int main(
         commandQueue.finish();
 
         auto start = std::chrono::system_clock::now();
-        for( int i = 0; i < iterations; i++ )
+        for( size_t i = 0; i < iterations; i++ )
         {
             commandQueue.enqueueNDRangeKernel(
                 kernel,
